+++ packages/core/src/node/runtimeModule/siteData.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import path from 'path';
 import {
   UserConfig,
@@ -394,6 +395,7 @@ export async function siteDataVMPlugin(context: FactoryContext) {
     title: userConfig?.title || '',
     description: userConfig?.description || '',
     icon: userConfig?.icon || '',
+    route: userConfig?.route,
     themeConfig: normalizeThemeConfig(userConfig, pages),
     base: userConfig?.base || '/',
     lang: userConfig?.lang || '',