+++ packages/core/src/runtime/Content.tsx
@@ -1,11 +1,26 @@
-import { ReactNode, Suspense } from 'react';
+import { ReactNode, Suspense, memo, ReactElement } from 'react';
 import { matchRoutes, useLocation } from 'react-router-dom';
+import siteData from 'virtual-site-data';
 import { normalizeRoutePath } from './utils';
-import { usePageData, useViewTransition } from './hooks';
+import { useViewTransition } from './hooks';
 
 const { routes } = process.env.__SSR__
   ? (require('virtual-routes-ssr') as typeof import('virtual-routes-ssr'))
   : (require('virtual-routes') as typeof import('virtual-routes'));
+
+function TransitionContentImpl(props: { el: ReactElement }) {
+  let element = props.el;
+  if (siteData?.themeConfig?.enableContentAnimation) {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    element = useViewTransition(props.el);
+  }
+  return element;
+}
+
+const TransitionContent = memo(
+  TransitionContentImpl,
+  (prevProps, nextProps) => prevProps.el === nextProps.el,
+);
 
 export const Content = ({ fallback = <></> }: { fallback?: ReactNode }) => {
   const { pathname } = useLocation();
@@ -14,17 +29,15 @@ export const Content = ({ fallback = <></> }: { fallback?: ReactNode }) => {
     return <div></div>;
   }
   const routesElement = matched[0].route.element;
-  /**
-   * useLayoutEffect to flush elements animation
-   */
-  let element = routesElement;
-  if (usePageData().siteData.themeConfig.enableContentAnimation) {
-    element = useViewTransition(routesElement);
-  }
 
   // React 17 Suspense SSR is not supported
   if (!process.env.__IS_REACT_18__ && process.env.__SSR__) {
-    return element;
+    return routesElement;
   }
-  return <Suspense fallback={fallback}>{element}</Suspense>;
+
+  return (
+    <Suspense fallback={fallback}>
+      <TransitionContent el={routesElement} />
+    </Suspense>
+  );
 };