+++ packages/core/src/theme-default/layout/DocLayout/docComponents/link.tsx
@@ -1,15 +1,21 @@
 import { ComponentProps } from 'react';
 import styles from './index.module.scss';
 import { usePathUtils } from '@/theme-default/logic';
+import { Link } from '@/theme-default/components/Link';
 
 export const A = (props: ComponentProps<'a'>) => {
-  const { href = '' } = props;
+  const { href = '', className = '' } = props;
   const { normalizeLinkHref } = usePathUtils();
-  return (
-    <a
-      {...props}
-      className={`${styles.link} ${props.className}`}
-      href={normalizeLinkHref(href)}
-    />
-  );
+
+  if (className.includes('header-anchor')) {
+    return <a {...props} className={`${styles.link} ${className}`} />;
+  } else {
+    return (
+      <Link
+        {...props}
+        className={`${className} ${styles.link} ${styles['inline-link']}`}
+        href={normalizeLinkHref(href)}
+      />
+    );
+  }
 };