+++ packages/core/src/theme-default/layout/HomeLayout/index.tsx
@@ -16,6 +16,7 @@ export function HomeLayout(props: HomeLayoutProps) {
   const {
     page: { frontmatter },
   } = usePageData();
+
   return (
     <div
       className="relative"