+++ packages/create-rspress/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-rspress
 
+## 1.0.0-beta.1
+
+## 1.0.0-beta.0
+
 ## 0.2.0
 
 ### Minor Changes